+++ seedfarmer/cli_groups/_remove_group.py
@@ -71,6 +71,18 @@ def remove() -> None:
     default=None,
 )
 @click.option(
+    "--profile",
+    default=None,
+    help="The AWS profile to use for boto3.Sessions",
+    required=False,
+)
+@click.option(
+    "--region",
+    default=None,
+    help="The AWS region of the toolchain",
+    required=False,
+)
+@click.option(
     "--target-account-id",
     default=None,
     help="Account Id to remove module data from, if specifed --target-region is required",
@@ -92,6 +104,8 @@ def remove_module_data(
     deployment: str,
     group: str,
     module: str,
+    profile: Optional[str],
+    region: Optional[str],
     project: Optional[str],
     target_account_id: Optional[str],
     target_region: Optional[str],
@@ -110,7 +124,7 @@ def remove_module_data(
     elif target_account_id is not None and target_region is not None:
         session = (
             SessionManager()
-            .get_or_create(project_name=project)
+            .get_or_create(project_name=project, profile=profile, region_name=region)
             .get_deployment_session(account_id=target_account_id, region_name=target_region)
         )